+++ index.html
@@ -9,6 +9,6 @@
 </head>
 <body>
     <div id="board"></div>
-    <script src="assets/scripts/app.js"></script>
+    <script type="module" src="assets/scripts/app.js"></script>
 </body>
 </html>
